add product loading skeleton
