--- app/dashboard/product/add/AddProductPage.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-type Props = {};
-
-const AddProductPage = (props: Props) => {
-  return <div>AddProductPage</div>;
-};
-
-export default AddProductPage;
